+++ src/backend/chat/tests/views/chat/conversations/test_conversation_with_history.py
@@ -234,6 +234,7 @@ def test_post_conversation_data_protocol_with_history(
     assert response_content == (
         '0:"Hello"\n'
         '0:" there"\n'
+        f'f:{{"messageId":"{mock_uuid4}"}}\n'
         'd:{"finishReason":"stop","usage":{"promptTokens":0,"completionTokens":0}}\n'
     )
 
@@ -410,6 +411,7 @@ def test_post_conversation_with_image_with_history(
     assert response_content == (
         '0:"I see a cat"\n'
         '0:" in the picture."\n'
+        f'f:{{"messageId":"{mock_uuid4}"}}\n'
         'd:{"finishReason":"stop","usage":{"promptTokens":0,"completionTokens":0}}\n'
     )
 
@@ -532,6 +534,7 @@ def test_post_conversation_tool_call_with_history(
         'a:{"toolCallId":"xLDcIljdsDrz0idal7tATWSMm2jhMj47","result":{"location":'
         '"Paris","temperature":22,"unit":"celsius"}}\n'
         '0:"The current weather in Paris is nice"\n'
+        f'f:{{"messageId":"{mock_uuid4}"}}\n'
         'd:{"finishReason":"stop","usage":{"promptTokens":0,"completionTokens":0}}\n'
     )
 
@@ -648,6 +651,7 @@ def test_post_conversation_tool_call_fails_with_history(
         'a:{"toolCallId":"xLDcIljdsDrz0idal7tATWSMm2jhMj47","result":"Unknown tool '
         "name: 'get_current_weather'. No tools available.\"}\n"
         '0:"I cannot give you an answer to that."\n'
+        f'f:{{"messageId":"{mock_uuid4}"}}\n'
         'd:{"finishReason":"stop","usage":{"promptTokens":0,"completionTokens":0}}\n'
     )
 
@@ -1183,6 +1187,7 @@ def test_post_conversation_with_existing_image_history(
     assert response_content == (
         '0:"Hello"\n'
         '0:" there"\n'
+        f'f:{{"messageId":"{mock_uuid4}"}}\n'
         'd:{"finishReason":"stop","usage":{"promptTokens":0,"completionTokens":0}}\n'
     )
 
@@ -1280,6 +1285,7 @@ def test_post_conversation_with_existing_tool_history(
         'a:{"toolCallId":"xLDcIljdsDrz0idal7tATWSMm2jhMj47","result":{"location":'
         '"Paris","temperature":22,"unit":"celsius"}}\n'
         '0:"The current weather in Paris is nice"\n'
+        f'f:{{"messageId":"{mock_uuid4}"}}\n'
         'd:{"finishReason":"stop","usage":{"promptTokens":0,"completionTokens":0}}\n'
     )
 
@@ -1462,6 +1468,7 @@ def test_post_conversation_add_image_to_conversation_with_tool_history(
     assert response_content == (
         '0:"I see a cat"\n'
         '0:" in the picture."\n'
+        f'f:{{"messageId":"{mock_uuid4}"}}\n'
         'd:{"finishReason":"stop","usage":{"promptTokens":0,"completionTokens":0}}\n'
     )